Revert some changes to avoid ABI breakage
