+++ sycl/include/CL/sycl.hpp
@@ -52,6 +52,7 @@
 #include <CL/sycl/sampler.hpp>
 #include <CL/sycl/specialization_id.hpp>
 #include <CL/sycl/stream.hpp>
+#include <CL/sycl/sub_group.hpp>
 #include <CL/sycl/types.hpp>
 #include <CL/sycl/usm.hpp>
 #include <CL/sycl/version.hpp>
